docs(ui): move non-config files out of .storybook directory
